% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{Airport_1}
\alias{Airport_1}
\alias{airport_fun_1}
\alias{airport_list_1}
\title{Airport spreadsheet}
\format{An object of class \code{list} of length 4.}
\usage{
airport_fun_1(global_level = as.data.frame(global_level_1),
  flight_level = flight_level_1, gate_level = gate_level_1,
  nat_level = nat_level_1)

airport_list_1
}
\description{
Airport_1
}
\keyword{datasets}
