% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research.R
\name{convert_stamps_to_hist}
\alias{convert_stamps_to_hist}
\title{Convert timestamps to binned data}
\usage{
convert_stamps_to_hist(x, breaks, ...)
}
\arguments{
\item{x}{object returned by airport_simulate()}

\item{breaks}{break points of histogram}
}
\description{
Convert timestamps to binned data
}
