% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\name{Global}
\alias{Global}
\alias{global_fun_1}
\alias{Global_fun}
\title{Global}
\usage{
global_fun_1(distance_dpl = 100, distance_imm = 150, distance_bh = 80,
  mu = 0.02, vm2 = 0.64, lag = 0.005, rate_bags = 4, nu_bags = 1.2,
  server_bags = 8, server_cus = 10, rate_cus = 3)
}
\description{
Global
}
