% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research.R
\name{safe_hist}
\alias{safe_hist}
\title{Histogram without error message for data points outside range}
\usage{
safe_hist(breaks, ...)
}
\arguments{
\item{breaks}{break points}
}
\description{
Histogram without error message for data points outside range
}
